DDLS-306 refactor slack lambda
