+++ lambdas/functions/monitor_notify_lambda/app/templates/cloudwatch_business_failure.txt
