+++ lambdas/functions/monitor_notify_lambda/app/templates/cloudwatch_event.txt
