+++ lambdas/functions/monitor_notify_lambda/app/templates/github_actions.txt
