+++ lambdas/functions/monitor_notify_lambda/app/templates/github_actions_scheduled_task.txt
