+++ lambdas/functions/monitor_notify_lambda/app/templates/sns_alerts.txt
