+++ lambdas/functions/monitor_notify_lambda/requirements.txt
