+++ lambdas/functions/monitor_notify_lambda/test_payloads/alarm_payload.json
@@ -1,0 +1,10 @@
+{
+  "Records": [
+    {
+      "Sns": {
+        "Message": "{\"AlarmName\": \"5xx test alarm\", \"NewStateValue\": \"bad\", \"OldStateValue\": \"good\"}",
+        "TopicArn": "arn:foo:bar:eu-west-1:123456789"
+      }
+    }
+  ]
+}
