+++ lambdas/functions/monitor_notify_lambda/test_payloads/cloudwatch_payload.json
@@ -1,0 +1,19 @@
+{
+  "scheduled-event-detail": {
+    "job-name": "example_job_name",
+    "log-group": "/aws/lambda/example-log-group",
+    "log-entries": [
+      {
+        "log_title": "Example Log Title 1",
+        "status": "success",
+        "count": 2,
+        "description": "The job executed successfully."
+      }
+    ],
+    "search-timespan": "15 minutes",
+    "bank-holidays": "false",
+    "channel-identifier-absent": "channel_absent_notification",
+    "channel-identifier-success": "channel_success_notification",
+    "channel-identifier-failure": "channel_failure_notification"
+  }
+}
