+++ terraform/account/region/lambda_monitor_notify.tf
@@ -1,16 +1,16 @@
 locals {
-  slack_lambda_function_name = "slack-notifier"
+  monitor_notify_lambda_function_name = "monitor-notify"
 }
 
-# INFO - Lambda used to manage all our slack notifications
-resource "aws_lambda_function" "slack_lambda" {
-  filename      = data.archive_file.slack_zip.output_path
-  function_name = local.slack_lambda_function_name
-  role          = aws_iam_role.lambda_slack.arn
-  handler       = "slack.lambda_handler"
+# INFO - Lambda used to monitor for different types of events and to notify us via slack
+resource "aws_lambda_function" "monitor_notify_lambda" {
+  filename      = data.archive_file.monitor_notify_zip.output_path
+  function_name = local.monitor_notify_lambda_function_name
+  role          = aws_iam_role.lambda_monitor_notify.arn
+  handler       = "monitor_notify.lambda_handler"
   runtime       = "python3.11"
   layers        = [aws_lambda_layer_version.lambda_layer.arn]
-  depends_on    = [aws_cloudwatch_log_group.slack_lambda]
+  depends_on    = [aws_cloudwatch_log_group.monitor_notify_lambda]
   timeout       = 300
   environment {
     variables = {
@@ -21,26 +21,26 @@ resource "aws_lambda_function" "slack_lambda" {
     mode = "Active"
   }
 
-  source_code_hash = filebase64sha256(data.archive_file.slack_zip.output_path)
+  source_code_hash = filebase64sha256(data.archive_file.monitor_notify_zip.output_path)
   tags = merge(
     var.default_tags,
-    { Name = "slack-${var.account.name}" },
+    { Name = "monitor-notify-${var.account.name}" },
   )
 }
 
-resource "aws_cloudwatch_log_group" "slack_lambda" {
-  name              = "/aws/lambda/${local.slack_lambda_function_name}"
+resource "aws_cloudwatch_log_group" "monitor_notify_lambda" {
+  name              = "/aws/lambda/${local.monitor_notify_lambda_function_name}"
   retention_in_days = 14
   kms_key_id        = aws_kms_key.cloudwatch_logs.arn
   tags = merge(
     var.default_tags,
-    { Name = "${var.account.name}-slack-log-group" },
+    { Name = "${var.account.name}-monitor-notify-log-group" },
   )
 }
 
-resource "aws_iam_role" "lambda_slack" {
+resource "aws_iam_role" "lambda_monitor_notify" {
   assume_role_policy = data.aws_iam_policy_document.lambda_slack_policy.json
-  name               = "lambda_slack"
+  name               = "lambda_monitor_notify"
   tags               = var.default_tags
 }
 
@@ -56,19 +56,19 @@ data "aws_iam_policy_document" "lambda_slack_policy" {
   }
 }
 
-resource "aws_iam_role_policy" "lambda_slack" {
-  name   = "lambda-slack"
-  policy = data.aws_iam_policy_document.lambda_slack.json
-  role   = aws_iam_role.lambda_slack.id
+resource "aws_iam_role_policy" "lambda_monitor_notify" {
+  name   = "lambda-monitor_notify"
+  policy = data.aws_iam_policy_document.lambda_monitor_notify.json
+  role   = aws_iam_role.lambda_monitor_notify.id
 }
 
-data "aws_iam_policy_document" "lambda_slack" {
+data "aws_iam_policy_document" "lambda_monitor_notify" {
   statement {
     sid    = "allowLogging"
     effect = "Allow"
     resources = [
-      aws_cloudwatch_log_group.slack_lambda.arn,
-      "${aws_cloudwatch_log_group.slack_lambda.arn}:*"
+      aws_cloudwatch_log_group.monitor_notify_lambda.arn,
+      "${aws_cloudwatch_log_group.monitor_notify_lambda.arn}:*"
     ]
     actions = [
       "logs:CreateLogStream",
@@ -112,22 +112,22 @@ data "aws_iam_policy_document" "lambda_slack" {
   }
 }
 
-data "archive_file" "slack_zip" {
+data "archive_file" "monitor_notify_zip" {
   type        = "zip"
-  source_dir  = "../../lambdas/functions/slack_lambda/app"
-  output_path = "../../lambdas/functions/slack_lambda/slack.zip"
+  source_dir  = "../../lambdas/functions/monitor_notify_lambda/app"
+  output_path = "../../lambdas/functions/monitor_notify_lambda/monitor_notify.zip"
 }
 
-data "archive_file" "slack_layer" {
+data "archive_file" "monitor_notify_layer" {
   type        = "zip"
-  source_dir  = "../../lambdas/functions/slack_lambda/layers"
-  output_path = "../../lambdas/functions/slack_lambda/layer.zip"
+  source_dir  = "../../lambdas/functions/monitor_notify_lambda/layers"
+  output_path = "../../lambdas/functions/monitor_notify_lambda/layer.zip"
 }
 
 resource "aws_lambda_layer_version" "lambda_layer" {
-  filename         = data.archive_file.slack_layer.output_path
-  source_code_hash = data.archive_file.slack_layer.output_base64sha256
-  layer_name       = "slack_notify_layer"
+  filename         = data.archive_file.monitor_notify_layer.output_path
+  source_code_hash = data.archive_file.monitor_notify_layer.output_base64sha256
+  layer_name       = "monitor_notify_layer"
 
   compatible_runtimes = ["python3.11"]
 
@@ -141,12 +141,12 @@ resource "aws_lambda_layer_version" "lambda_layer" {
 resource "aws_lambda_permission" "sns" {
   statement_id  = "AllowExecutionFromSNSTopic"
   action        = "lambda:InvokeFunction"
-  function_name = aws_lambda_function.slack_lambda.function_name
+  function_name = aws_lambda_function.monitor_notify_lambda.function_name
   principal     = "sns.amazonaws.com"
   source_arn    = aws_sns_topic.alerts.arn
   lifecycle {
     replace_triggered_by = [
-      aws_lambda_function.slack_lambda
+      aws_lambda_function.monitor_notify_lambda
     ]
   }
 }
@@ -154,24 +154,24 @@ resource "aws_lambda_permission" "sns" {
 resource "aws_lambda_permission" "scheduled_checks" {
   statement_id  = "AllowExecutionFromScheduledCheck"
   action        = "lambda:InvokeFunction"
-  function_name = aws_lambda_function.slack_lambda.function_name
+  function_name = aws_lambda_function.monitor_notify_lambda.function_name
   principal     = "events.amazonaws.com"
   source_arn    = "arn:aws:events:${data.aws_region.current.name}:${data.aws_caller_identity.current.account_id}:rule/check-*"
   lifecycle {
     replace_triggered_by = [
-      aws_lambda_function.slack_lambda
+      aws_lambda_function.monitor_notify_lambda
     ]
   }
 }
 
 resource "aws_lambda_permission" "sns_availability" {
   action        = "lambda:InvokeFunction"
-  function_name = aws_lambda_function.slack_lambda.function_name
+  function_name = aws_lambda_function.monitor_notify_lambda.function_name
   principal     = "sns.amazonaws.com"
   source_arn    = aws_sns_topic.availability-alert.arn
   lifecycle {
     replace_triggered_by = [
-      aws_lambda_function.slack_lambda
+      aws_lambda_function.monitor_notify_lambda
     ]
   }
 }
